implementing guards on routes for authentication
