+++ src/app.controller.ts
@@ -1,26 +1,8 @@
-import { Controller, Get, Post, Request, UseGuards } from '@nestjs/common';
-import { AuthGuard } from '@nestjs/passport';
-import { LocalAuthGuard } from './auth/local-auth.guard';
-import { AuthService } from './auth/auth.service';
-import { JwtAuthGuard } from './auth/jwt-auth.guard';
+import { Controller, Get } from '@nestjs/common';
 
 @Controller()
 export class AppController {
-  constructor(private authService: AuthService) {}
-
-  @UseGuards(LocalAuthGuard)
-  @Post('auth/login')
-  async login(@Request() req){
-    console.log(req);
-    return req.user;
-  }
-
-  //@UseGuards(JwtAuthGuard)
-  @Get('/profile')
-  getProfile(@Request() req){
-    console.log('returned: ' + req.user);
-    return req.user;
-  }
+  constructor() {}
 
   @Get()
   getHello(): string {
