+++ src/auth/jwt.strategy.ts
@@ -9,7 +9,8 @@ export class JwtStrategy extends PassportStrategy(Strategy){
         super({
             jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
             ignoreExpiration: false,
-            secretOrKey: process.env.SECRET
+            //secretOrKey: process.env.SECRET
+            secretOrKey: jwtConstants.secret
         });
     }
