+++ src/auth/local.strategy.ts
@@ -3,14 +3,13 @@ import { PassportStrategy } from '@nestjs/passport';
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { jwtConstants } from './constants';
-import { stringify } from 'querystring';
 
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy) {
   constructor(private authService: AuthService) {
     super({
       usernameField: 'email',
-      passwordFiled: 'password'
+      passwordField: 'password'
     });
   }
