+++ src/modules/calendar/calendar.controller.ts
@@ -1,4 +1,4 @@
-import { Body, ClassSerializerInterceptor, Controller, Delete, HttpCode, HttpException, HttpStatus, Param, Patch, Post, UseInterceptors } from '@nestjs/common';
+import { Body, ClassSerializerInterceptor, Controller, Delete, HttpCode, HttpException, HttpStatus, Param, Patch, Post, UseGuards, UseInterceptors } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { ApiOperation, ApiBody, ApiParam, ApiOkResponse, ApiBadRequestResponse, ApiUnauthorizedResponse, ApiNotFoundResponse, ApiTags } from '@nestjs/swagger';
 import { model, Model, mongo, Types, Schema } from 'mongoose';
@@ -6,6 +6,7 @@ import { User, UserDocument } from 'src/entities/user/user';
 import { Calendar, CalendarDocument } from 'src/entities/calendar/calendar';
 import { CalendarCreateDTO } from 'src/entities/calendar/create-calendar';
 import { CalendarEditDTO } from 'src/entities/calendar/edit-calendar';
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
 
 
 @Controller()
@@ -26,6 +27,7 @@ export class CalendarController {
   @ApiNotFoundResponse({ description: 'User does not exist.'})
   @ApiTags('Calendar')
   @UseInterceptors(ClassSerializerInterceptor)
+  @UseGuards(JwtAuthGuard)
   async createCalendar(@Param('userId') id: string, @Body() calendar: CalendarCreateDTO): Promise<string>{ // function used to return Promise<Calendar>
     
     if(id === null){
@@ -68,10 +70,12 @@ export class CalendarController {
   @ApiUnauthorizedResponse({ description: 'User does not have access.'})
   @ApiNotFoundResponse({ description: ''})
   @ApiTags('Calendar')
+  @UseGuards(JwtAuthGuard)
   async updateCalendar(
     @Param('userId') id: string, 
     @Param('calendarId') calendarId: string,
-    @Body() calendar: CalendarEditDTO): Promise<string> {
+    @Body() calendar: CalendarEditDTO
+    ): Promise<string> {
 
 
     if(calendar == null){
